+++ Devcade2048/App/Render/Animation/FromTabAnimationState.cs
@@ -27,7 +27,7 @@ public class FromTabAnimationState : TransientAnimationState {
         // This backgroundColor pixel is stretched to cover everything. 
         Rectangle screen = new Rectangle(0, 0, gd.DisplayMode.Width, gd.DisplayMode.Height);
 
-        DrawAsset(uiCover, screen, Color.White * (float) (FastStart()));
+        DrawAsset(uiCover, screen, Color.White * FastStart());
         base.Draw();
     }
 }
