+++ Devcade2048/App/Render/Animation/ResetWinAnimationState.cs
@@ -1,6 +1,5 @@
 using System;
 using Microsoft.Xna.Framework;
-using Microsoft.Xna.Framework.Graphics;
 
 namespace Devcade2048.App.Render.Animation;
 
@@ -29,7 +28,7 @@ public class ResetWinAnimationState : TransientAnimationState {
     }
 
     private void DrawScore() {
-        float opacity = 4 * MathF.Abs((float) PercentComplete() - 0.75f);
+        float opacity = 4 * MathF.Abs(PercentComplete() - 0.75f);
         if (opacity > 1) opacity = 1;
         Color scoreColor = Color.Black * opacity;
 
@@ -57,7 +56,7 @@ public class ResetWinAnimationState : TransientAnimationState {
     }
 
     public void DrawWinText() {
-        float opacity = (float) (1 - FastStart());
+        float opacity = 1 - FastStart();
         DrawAsset(Asset.BigFont, "YOU WIN!", new Vector2(20, 700), Color.Black * opacity);
         DrawAsset(Asset.Button, new Vector2(20, 720), Color.Red * opacity);
         DrawAsset(Asset.BigFont, "Play again", new Vector2(125, 750), Color.Red * opacity);
